Integral types in BasicTypes.h are used from boost/cstdint.hpp. Integral types limit macros replaced with numeric_limits struct. float32 and float64 marked deprecated. Added ResourceTypes.cpp to CMakeLists.txt.

+++ src/Editor/EditorGUI.cpp
@@ -158,7 +158,7 @@ void EditorGUI::UpdateEntityEditorWindow()
 			mEntityEditorLayout = new GUISystem::VerticalLayout(entityEditorPane, entityEditorContentPane);
 
 		// Clear all the content of Entity Editor.
-		int childCount = entityEditorContentPane->getChildCount();
+		size_t childCount = entityEditorContentPane->getChildCount();
 		for (int i = (childCount - 1); i >= 0; --i)
 		{
 			gCEGUIWM.destroyWindow(entityEditorContentPane->getChildAtIdx(i));
