+++ src/Editor/ResourceWindow.cpp
@@ -28,7 +28,7 @@ void Editor::ResourceWindow::Init()
 	gGUIMgr.GetRootLayout()->addChildWindow(w);
 	CEGUI::ItemListbox* list = static_cast<CEGUI::ItemListbox*>(w->getChild(w->getName() + "/List"));
 	OC_ASSERT(list != 0);
-	for (uint32 i = 0; i < list->getContentPane()->getChildCount(); ++i)
+	for (size_t i = 0; i < list->getContentPane()->getChildCount(); ++i)
 	{
 		CEGUI::Window* aa = list->getContentPane()->getChildAtIdx(i);
 		ocWarning << aa->getName();
@@ -50,7 +50,7 @@ void Editor::ResourceWindow::Init()
 
 		uint32 depth = 0;
 		size_t searchFrom = 0;
-		size_t match = 0;		
+		size_t match = 0;
 		do
 		{
 			match = path.find('/', searchFrom);
